apps/qrb/requirements.txt
python-dotenv
gunicorn
jsonpath
flask
mongoengine
PyJWT
pyzbar
Pillow
python-dateutil
cryptography